# src/AirlineService/swagger.json
{
   "openapi":"3.0.1",
   "info":{
      "description":"A description about all callable AirlineService API.",
      "version":"1.0.5",
      "title":"AirlineService API"
   },
   "tags":[
   ],
   "servers":[
      {
         "url":"http://localhost:8060",
         "description":""
      }
   ],
   "paths":{
      "/getFlights":{
         "post":{
            "tags":[
               "flight-offer-controller"
            ],
            "operationId":"getOffers",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "type":"array",
                        "items":{
                           "$ref":"#/components/schemas/UserRequest"
                        }
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/Flight"
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "/getTickets":{
         "get":{
            "tags":[
               "flight-offer-controller"
            ],
            "operationId":"sendPdfFiles",
            "parameters":[
               {
                  "name":"id",
                  "in":"query",
                  "required":true,
                  "schema":{
                     "type":"array",
                     "items":{
                        "type":"integer",
                        "format":"int64"
                     }
                  }
               }
            ],
            "responses":{
               "200":{
                 "description": "Ticket file example",
                  "content":{
                    "application/pdf":{
                      "schema":{
                        "type": "string",
                        "format": "binary",
                          "sampleFile":{
                          "externalValue":"https://github.com/Vallasc/ACMEsky/tree/main/docs/swagger-ui/ticketSample.pdf"
                          }
            
                      }
                    }
                  }
              }
          }
        }
      },
      "/getIban":{
         "get":{
            "tags":[
               "flight-offer-controller"
            ],
            "operationId":"sendIban",
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/Iban"
                        }
                     }
                  }
               }
            }
         }
      },
      "/getAll":{
         "get":{
            "tags":[
               "flight-offer-controller"
            ],
            "operationId":"getAll",
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/FlightOffer"
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "/notPurchasedOffer":{
         "post":{
            "tags":[
               "flight-offer-controller"
            ],
            "operationId":"deleteOffer",
            "parameters":[
               {
                  "name":"id",
                  "in":"query",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "format":"int64"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "type":"string",
                           "example":"Ok"
                        }
                     }
                  }
               }
        
            }
         }
      }
   },
   "components":{
      "schemas":{
         "UserRequest":{
            "required":[
               "departure",
               "arrival",
               "departureDate"
            ],
            "type":"object",
            "properties":{
               "departure":{
                  "type":"string",
                  "example":"AMS"
               },
               "arrival":{
                  "type":"string",
                  "example":"BOS"
               },
               "departureDate":{
                  "type":"string",
                  "format":"date-time",
                  "example":"2021-08-27T20:00:00+01:00"
               }
            }
         },
         "Flight":{
            "type":"object",
            "properties":{
               "id":{
                  "type":"integer",
                  "format":"int64"
               },
               "departureId":{
                  "type":"string",
                  "example":"2218"
                  
               },
               "departureTime":{
                  "type":"string",
                  "format":"date-time",
                  "example":"2021-08-27T19:00:00Z"
               },
               "arrivalTime":{
                  "type":"string",
                  "format":"date-time",
                  "example":"2021-08-27T20:00:00Z"
               },
               "arrivalId":{
                  "type":"string",
                  "example":"2890"
               },
               "airline_id":{
                  "type":"string",
                  "example":"1"
               },
               "price":{
                  "type":"number",
                  "format":"double",
                  "example":130.00
               },
               "place":{
                  "type":"string",
                  "example":"25"
               }
            }
         },
         "Iban":{
            "type":"object",
            "properties":{
               "iban":{
                  "type":"string",
                  "example":"IT44A0300203280446531574273"
               }
            }
         },
         "FlightOffer":{
            "type":"object",
            "properties":{
               "id":{
                  "type":"integer",
                  "format":"int64"
               },
               "departureId":{
                  "type":"string",
                  "example":"2218"
                  
               },
               "departureTime":{
                  "type":"string",
                  "format":"date-time",
                  "example":"2021-08-27T19:00:00Z"
               },
               "arrivalTime":{
                  "type":"string",
                  "format":"date-time",
                  "example":"2021-08-27T20:00:00Z"
               },
               "arrivalId":{
                  "type":"string",
                  "example":"2890"
               },
               "airline_id":{
                  "type":"string",
                  "example":"1"
               },
               "price":{
                  "type":"number",
                  "format":"double",
                  "example":130.00
               },
               "place":{
                  "type":"string",
                  "example":"25"
               },
               "expiryDate":{
                  "type":"string",
                  "format":"date-time",
                  "example":"2021-08-20T19:00:00Z"
               },
               "soldFlag":{
                  "type":"boolean"
               }
            }
         }
      }
   }
}